add core_environment.cpp
