+++ .gitignore
@@ -22,3 +22,7 @@ project/addons/*/bin
 *.includes
 *.idb
 *.exp
+
+# ide
+.idea
+.vscode
