+++ src/loader.cpp
@@ -87,93 +87,6 @@ static void audio_init(int frequency) {
 
 static void audio_deinit() {}
 
-static void core_log(enum retro_log_level level, const char * fmt, ...) {
-  char buffer[4096] = {0};
-  static const char * levelstr[] = {
-    "dbg",
-    "inf",
-    "wrn",
-    "err"
-  };
-  va_list va;
-
-  va_start(va, fmt);
-  vsnprintf(buffer, sizeof(buffer), fmt, va);
-  va_end(va);
-
-  if (level == 0)
-    return;
-
-  fprintf(stderr, "[%s] %s", levelstr[level], buffer);
-  fflush(stderr);
-
-  if (level == RETRO_LOG_ERROR) {
-    exit(EXIT_FAILURE);
-  }
-}
-
-static bool core_environment(unsigned cmd, void * data) {
-  switch (cmd) {
-    case RETRO_ENVIRONMENT_GET_LOG_INTERFACE: {
-      struct retro_log_callback* cb = (struct retro_log_callback*)data;
-      cb->log = core_log;
-    }
-    break;
-
-    case RETRO_ENVIRONMENT_GET_CAN_DUPE: {
-      bool* bval = (bool*)data; // NOLINT
-      *bval = true;
-    }
-    break;
-
-    // case RETRO_ENVIRONMENT_SET_PIXEL_FORMAT: {
-    //   core_log(RETRO_LOG_DEBUG, "[gdretro] RETRO_ENVIRONMENT_SET_PIXEL_FORMAT");
-    //   return true;
-    // }
-
-    case RETRO_ENVIRONMENT_SET_PIXEL_FORMAT:
-    {
-        const enum retro_pixel_format *fmt = (enum retro_pixel_format *)data;
-
-        if ( *fmt > RETRO_PIXEL_FORMAT_RGB565 )
-        {
-            return false;
-        }
-
-        core_log(RETRO_LOG_DEBUG, "[gdretro] Core setting pixel format" );
-        return GDRetro::get_singleton()->core_video_set_pixel_format( *fmt );
-    }
-
-    case RETRO_ENVIRONMENT_GET_SYSTEM_DIRECTORY:
-    case RETRO_ENVIRONMENT_GET_SAVE_DIRECTORY:
-    case RETRO_ENVIRONMENT_GET_CONTENT_DIRECTORY:
-    case RETRO_ENVIRONMENT_GET_LIBRETRO_PATH: {
-      *(const char**)data = ".";
-      return true;
-    }
-
-    case RETRO_ENVIRONMENT_SET_MESSAGE: {
-      const struct retro_message* message = (const struct retro_message*)data;
-      core_log(
-        RETRO_LOG_DEBUG,
-        "[gdretro] RETRO_ENVIRONMENT_SET_MESSAGE: %s\n",
-        message->msg);
-      break;
-    }
-
-    case RETRO_ENVIRONMENT_SHUTDOWN: {
-      core_log(RETRO_LOG_DEBUG, "[gdretro] RETRO_ENVIRONMENT_SHUTDOWN");
-      break;
-    }
-
-    default: {
-      core_log(RETRO_LOG_DEBUG, "[gdretro] Unhandled env #%u", cmd);
-      return false;
-    }
-  }
-
-  return true;
-}
 
 static void core_input_poll(void) {
   // Nothing
@@ -239,12 +152,13 @@ bool core_load(const char * sofile) {
   load_sym(set_audio_sample_batch, retro_set_audio_sample_batch);
 
  
-  set_environment(core_environment);
+  set_environment([]( unsigned cmd, void *data ) { 
+    return GDRetro::get_singleton()->core_environment( cmd, data ); } 
+  );
 
   set_video_refresh( []( const void *data, unsigned width, unsigned height, size_t pitch ) {
-    // MINE
     GDRetro::get_singleton()->core_video_refresh( data, width, height, pitch );
- } );
+  });
 
   set_input_poll(core_input_poll);
   set_input_state(core_input_state);
